refactor ui work to a new class
